drm-extras: replace edid-rs with libdisplay-info
