+++ smithay-drm-extras/Cargo.toml
@@ -6,15 +6,12 @@ license = "MIT"
 authors = ["Bartłomiej Maryńczak <marynczakbartlomiej@gmail.com>"]
 
 [dependencies]
-edid-rs = "0.1.0"
+libdisplay-info = { version = "0.1.0", optional = true }
 drm = { version = "0.12.0" }
 
 [features]
-default = []
-generate-hwdata = ["pkg-config"]
+default = ["display-info"]
+display-info = ["libdisplay-info"]
 
 [dev-dependencies.smithay]
 path = "../"
-
-[build-dependencies]
-pkg-config = { version = "0.3.26", optional = true }
