+++ smithay-drm-extras/src/lib.rs
@@ -2,15 +2,15 @@
 //!
 //! This crate contains some extra abstractions and helpers over DRM
 //!
-//! - [`edid`] is responsible for extraction of information from DRM connectors
+//! - [`display_info`] is responsible for extraction of information from DRM connectors
 //! - [`drm_scanner`] is responsible for detecting connector connected and
 //!   disconnected events, as well as mapping CRTC to them.
 //!
 //! ### Features
-//! - `generate-hwdata` - If enabled [hwdata](https://github.com/vcrhonek/hwdata) code will be regenerated using `hwdata` system package
+//! - `display_info` - If enabled `display_info` functionality is enabled through `libdisplay-info` integration
 
 #![warn(missing_docs, missing_debug_implementations)]
 
+#[cfg(feature = "display-info")]
+pub mod display_info;
 pub mod drm_scanner;
-pub mod edid;
-mod hwdata;
